chore(gate): raise main.rs's ceiling to 1548 — the last red gate on main
#937 landed, independently, everything the earlier commits on this branch
fixed: the `ProofStep`/`ProofTree` import #908's merge dropped, the retired
`palette::EMBER` references, `connect_cmd`'s `Zeroizing` import and its
dependency, the private-intra-doc-link on `IndexStats`, and rustfmt over the
seven unformatted files. Keeping this branch's versions of those was worse
than redundant — a second `use zeroize::Zeroizing;` is E0252, and the
`config.rs` split put INTERACTIVE_CREDENTIALS out of reach of a test #937
added that reads the static directly. All of it is reverted to main's
versions, so this branch's content delta is one line.

That line is the only gate still red on main. `check-file-size.sh` is the sole
failing step of main's own CI run (30512789019); fmt, clippy, doc, test, the
release smoke build and cargo deny/audit all passed. #937 added a
`Completions` subcommand declaration and `long_version_static()` to
stella-cli/src/main.rs, taking it to 1548 against a 1545 ceiling — precisely
the "subcommand or module declaration in an already-oversized lib.rs/main.rs"
growth the gate's own message calls irreducible and answers with a recorded
baseline bump.

Hand-edited rather than regenerated: `--update` re-sorts the file under this
shell's locale collation, which swaps agent.rs/agent_tests.rs and adds
cosmetic churn to a review. The gate reads the baseline into a map, so
ordering is not load-bearing. One line changed, no entry added or removed:
37 grandfathered files before and after.
