+++ scripts/file-size-baseline.txt
@@ -20,7 +20,7 @@
 1576 stella-cli/src/candidate_ws.rs
 4570 stella-cli/src/command_deck.rs
 1518 stella-cli/src/config.rs
-1545 stella-cli/src/main.rs
+1548 stella-cli/src/main.rs
 2095 stella-core/src/bus.rs
 2138 stella-core/src/driver.rs
 2820 stella-core/src/driver/tests.rs
